requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.7.0
transformers==4.36.0
torch==2.1.0
huggingface_hub==0.19.4
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
groq==0.4.1
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
tokenizers==0.15.0